plugins.xunit: Convert size to int
On python3 divisions are converted to float automatically, but floats
can't be supplied to seek.

+++ avocado/plugins/xunit.py
@@ -71,7 +71,7 @@ class XUnitResult(Result):
                     if log_size < max_log_size:
                         text_output = logfile_obj.read()
                     else:
-                        size = max_log_size / 2
+                        size = int(max_log_size / 2)
                         logfile_obj.seek(0, 0)
                         text_output = logfile_obj.read(size)
                         text_output += ("\n\n--[ CUT DUE TO XML PER TEST "
